refactor: extract indexer/cache.rs (phase 3, step 8)
Move FileCacheEntry, IndexCache, xdg_cache_base, status_cache_path,
write_status_file, workspace_cache_path, try_load_cache,
cache_entry_to_file_result, save_cache out of indexer.rs into
src/indexer/cache.rs. save_cache_to_disk becomes a thin 3-line wrapper.

Add cache_tests.rs with 6 tests: supertypes extraction, hash preservation,
path stability/differentiation, missing-file returns None, save/load roundtrip.
