+++ src/indexer.rs
@@ -5,7 +5,6 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, RwLock};
 
 use dashmap::{DashMap, DashSet};
-use serde::{Deserialize, Serialize};
 use tower_lsp::lsp_types::*;
 
 // ─── LSP progress notification helper ────────────────────────────────────────
@@ -65,6 +64,10 @@ pub(crate) use self::infer::{
     find_last_dot_at_depth_zero,
 };
 
+mod cache;
+use self::cache::{IndexCache, try_load_cache, write_status_file, save_cache, FileCacheEntry, CACHE_VERSION, cache_entry_to_file_result};
+pub(crate) use self::cache::workspace_cache_path;
+
 // ─── RAII guard for indexing_in_progress flag ─────────────────────────────────
 
 /// RAII guard that clears `indexing_in_progress` on drop (success, panic, or early return).
@@ -77,28 +80,6 @@ impl Drop for IndexingGuard {
         self.indexer.indexing_in_progress.store(false, std::sync::atomic::Ordering::Release);
         log::debug!("IndexingGuard: cleared indexing_in_progress flag");
     }
-}
-
-// ─── Indexing status file ─────────────────────────────────────────────────────
-
-/// Human-readable status written to `~/.cache/kotlin-lsp/status.json`.
-/// The skill extension reads this to report loading state with time estimates.
-fn status_cache_path() -> PathBuf {
-    let cache_base = std::env::var("XDG_CACHE_HOME")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| {
-            let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
-            PathBuf::from(home).join(".cache")
-        });
-    cache_base.join("kotlin-lsp").join("status.json")
-}
-
-fn write_status_file(content: &str) {
-    let path = status_cache_path();
-    if let Some(parent) = path.parent() {
-        let _ = std::fs::create_dir_all(parent);
-    }
-    let _ = std::fs::write(&path, content);
 }
 
 /// Hard cap on workspace files indexed eagerly in LSP mode.
@@ -122,89 +103,6 @@ pub fn resolve_max_files(default: usize) -> usize {
         .ok()
         .and_then(|v| v.parse().ok())
         .unwrap_or(default)
-}
-
-// ─── Disk cache ──────────────────────────────────────────────────────────────
-
-/// Bump when the serialized format changes; invalidates any older cache files.
-const CACHE_VERSION: u32 = 4;
-
-/// Per-file entry stored in the on-disk index cache.
-#[derive(Serialize, Deserialize)]
-pub(crate) struct FileCacheEntry {
-    /// File mtime (seconds since Unix epoch) — primary cache validity check.
-    mtime_secs: u64,
-    /// File size in bytes — secondary guard for same-second edits (1s mtime resolution).
-    file_size: u64,
-    /// FNV-1a content hash — tertiary guard for mtime collisions / FAT FS.
-    content_hash: u64,
-    /// Parsed symbol data for this file.
-    file_data: FileData,
-}
-
-/// Complete serialized index, written to `~/.cache/kotlin-lsp/<root-hash>/index.bin`.
-#[derive(Serialize, Deserialize)]
-struct IndexCache {
-    version: u32,
-    /// True when this cache was built from a complete (non-truncated) workspace scan.
-    /// Only set to true when `total <= max` at index time.
-    /// When false, the entries may be a partial subset of the workspace — warm-manifest
-    /// mode is disabled to avoid hiding files that were never indexed.
-    #[serde(default)]
-    complete_scan: bool,
-    /// Absolute path string → per-file cached data.
-    entries: HashMap<String, FileCacheEntry>,
-}
-
-/// Returns the cache file path for the given workspace root.
-pub(crate) fn workspace_cache_path(root: &Path) -> PathBuf {
-    // Use canonicalized absolute path so equivalent roots map to same cache.
-    let canonical = std::fs::canonicalize(root).unwrap_or_else(|_| root.to_path_buf());
-    // Hash canonical path for filesystem-friendly cache directory name.
-    let root_hash = {
-        use sha2::{Digest, Sha256};
-        let mut hasher = Sha256::new();
-        hasher.update(canonical.to_string_lossy().as_bytes());
-        let digest = hasher.finalize();
-        // take first 8 bytes as u64
-        let mut bytes = [0u8; 8];
-        bytes.copy_from_slice(&digest[..8]);
-        u64::from_be_bytes(bytes)
-    };
-    let cache_base = std::env::var("XDG_CACHE_HOME")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| {
-            let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
-            PathBuf::from(home).join(".cache")
-        });
-    cache_base
-        .join("kotlin-lsp")
-        .join(format!("{root_hash:016x}"))
-        .join("index.bin")
-}
-
-/// Load and validate the on-disk cache.  Returns `None` if absent / stale / corrupt.
-fn try_load_cache(root: &Path) -> Option<IndexCache> {
-    let path = workspace_cache_path(root);
-    let bytes = std::fs::read(&path).ok()?;
-    let cache: IndexCache = match bincode::deserialize(&bytes) {
-        Ok(c) => c,
-        Err(e) => {
-            log::warn!("Cache deserialize failed (struct layout changed?): {e} — will re-index. \
-                        Delete {path} to suppress this warning.", path = path.display());
-            return None;
-        }
-    };
-    if cache.version != CACHE_VERSION {
-        log::info!("Cache version mismatch — will re-index");
-        return None;
-    }
-    log::info!(
-        "Loaded index cache ({} files) from {}",
-        cache.entries.len(),
-        path.display()
-    );
-    Some(cache)
 }
 
 // ─── Pure helper types ────────────────────────────────────────────────────────
@@ -299,39 +197,6 @@ pub(crate) fn stale_keys_for(uri: &Url, old_data: &crate::types::FileData) -> St
         definition_names,
         qualified_keys,
         package: old_data.package.clone(),
-    }
-}
-
-/// Pure: convert a disk-cache entry into a `FileIndexResult` ready for indexing.
-/// Reconstructs `supertypes` from cached lines (not stored in cache) so that
-/// `goToImplementation` works correctly on cache hits.
-pub(crate) fn cache_entry_to_file_result(uri: &Url, entry: &FileCacheEntry) -> FileIndexResult {
-    let data = &entry.file_data;
-    let class_kinds = [
-        SymbolKind::CLASS, SymbolKind::INTERFACE, SymbolKind::STRUCT,
-        SymbolKind::ENUM, SymbolKind::OBJECT,
-    ];
-    let mut supertypes: Vec<(String, Location)> = Vec::new();
-    for sym in &data.symbols {
-        if !class_kinds.contains(&sym.kind) { continue; }
-        let start = sym.selection_range.start.line as usize;
-        let limit  = (start + 10).min(data.lines.len());
-        let mut decl_lines: Vec<String> = Vec::new();
-        for line in &data.lines[start..limit] {
-            decl_lines.push(line.clone());
-            if line.contains('{') { break; }
-        }
-        let class_loc = Location { uri: uri.clone(), range: sym.selection_range };
-        for super_name in crate::resolver::extract_supers_from_lines(&decl_lines) {
-            supertypes.push((super_name, class_loc.clone()));
-        }
-    }
-    FileIndexResult {
-        uri: uri.clone(),
-        data: data.clone(),
-        supertypes,
-        content_hash: entry.content_hash,
-        error: None,
     }
 }
 
@@ -1015,67 +880,8 @@ impl Indexer {
             Some(r) => r,
             None    => return,
         };
-        let cache_path = workspace_cache_path(root);
-        if let Some(parent) = cache_path.parent() {
-            if let Err(e) = std::fs::create_dir_all(parent) {
-                log::warn!("Cache: could not create directory: {e}");
-                return;
-            }
-        }
-
         let complete_scan = self.last_scan_complete.load(std::sync::atomic::Ordering::Acquire);
-        let mut entries: HashMap<String, FileCacheEntry> = HashMap::new();
-        for file_ref in &self.files {
-            let uri_str = file_ref.key();
-            // Skip library-source files — they are re-indexed from sourcePaths on each startup.
-            if self.library_uris.contains(uri_str) { continue; }
-            let data    = file_ref.value();
-            let hash    = self.content_hashes.get(uri_str).map(|h| *h).unwrap_or(0);
-            if let Ok(url) = uri_str.parse::<Url>() {
-                if let Ok(path) = url.to_file_path() {
-                    let meta = std::fs::metadata(&path);
-                    let mtime = meta.as_ref().ok()
-                        .and_then(|m| m.modified().ok())
-                        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-                        .map(|d| d.as_secs())
-                        .unwrap_or(0);
-                    let file_size = meta.as_ref().map(|m| m.len()).unwrap_or(0);
-                    entries.insert(
-                        path.to_string_lossy().to_string(),
-                        FileCacheEntry { mtime_secs: mtime, file_size, content_hash: hash, file_data: (**data).clone() },
-                    );
-                }
-            }
-        }
-
-        let cache = IndexCache { version: CACHE_VERSION, complete_scan, entries };
-        match bincode::serialize(&cache) {
-            Ok(bytes) => {
-                // Don't overwrite a complete cache with an incomplete one.
-                // This prevents editor servers (which may load only part of the
-                // workspace) from truncating a cache built by --index-only.
-                if !complete_scan {
-                    if let Ok(meta) = std::fs::metadata(&cache_path) {
-                        if meta.len() > bytes.len() as u64 {
-                            log::info!(
-                                "Cache save skipped: existing cache ({} KB) is larger than \
-                                 incomplete new cache ({} KB)",
-                                meta.len() / 1024, bytes.len() / 1024
-                            );
-                            return;
-                        }
-                    }
-                }
-                match std::fs::write(&cache_path, &bytes) {
-                    Ok(()) => log::info!(
-                        "Cache saved ({} files, {} KB) → {}",
-                        cache.entries.len(), bytes.len() / 1024, cache_path.display()
-                    ),
-                    Err(e) => log::warn!("Cache write failed: {e}"),
-                }
-            }
-            Err(e) => log::warn!("Cache serialize failed: {e}"),
-        }
+        save_cache(root, &self.files, &self.content_hashes, &self.library_uris, complete_scan);
     }
 
     // ────────────────────────────────────────────────────────────────────────
@@ -4653,40 +4459,6 @@ internal class ContactAddressInteractor @Inject constructor(
         let result = super::resolve_max_files(2000);
         std::env::remove_var("KOTLIN_LSP_MAX_FILES");
         assert_eq!(result, 2000);
-    }
-
-    // ── cache_entry_to_file_result ────────────────────────────────────────────
-
-    /// cache_entry_to_file_result must reconstruct supertypes from FileData.lines
-    /// even when the FileCacheEntry was loaded from disk (lines are always cached).
-    #[test]
-    fn cache_entry_to_file_result_supertypes_extracted() {
-        let u = uri("/Cat.kt");
-        let mut data = crate::types::FileData::default();
-        data.lines = Arc::new(vec![
-            "class Cat : IAnimal {".into(),
-            "    fun meow() {}".into(),
-            "}".into(),
-        ]);
-        data.symbols.push(crate::types::SymbolEntry {
-            name: "Cat".into(),
-            kind: tower_lsp::lsp_types::SymbolKind::CLASS,
-            visibility: crate::types::Visibility::Public,
-            range: Default::default(),
-            selection_range: Default::default(),
-            detail: String::new(),
-        });
-
-        let entry = FileCacheEntry {
-            mtime_secs: 100,
-            file_size: 0,
-            content_hash: 42,
-            file_data: data,
-        };
-
-        let result = super::cache_entry_to_file_result(&u, &entry);
-        let super_names: Vec<&str> = result.supertypes.iter().map(|(n, _)| n.as_str()).collect();
-        assert!(super_names.contains(&"IAnimal"), "IAnimal missing from supertypes: {super_names:?}");
     }
 
     // ── apply_workspace_result ────────────────────────────────────────────────
